+++ html/history.md
@@ -1,13 +1,16 @@
-# History
+# WCA Regulations and Guidelines
 
-## Older versions of the Regulations
+## Older Versions of the Regulations
 
 Until 2011, the Regulations were maintained by Ron van Bruchem and the WCA Board. Since then, the [WCA Regulations Committee](http://www.worldcubeassociation.org/contact/wrc) is in charge of them.  
 For the 2013 release, the Regulations were split into two documents: the Regulations and the Guidelines.
 
 Here are older versions regulations. None of them are official anymore.
 
+- [2012: No new Regulations]
+- [2011: No new Regulations]
 - [2010](regulations2010.html) ([Changes](regulations_history2010.html))
+- [2009: No new Regulations]
 - [2008](regulations2008.html) ([Changes](regulations_history2009.html))
 - [2007](regulations2007.html) ([Changes](regulations_history2008.html))
 - [2006 v2](regulations2006v2.html)